refactor: SlidingPiece 패키지 이동
