+++ src/main/java/domain/piece/sliding/Queen.java
@@ -2,7 +2,6 @@ package domain.piece.sliding;
 
 import domain.Direction;
 import domain.Team;
-import domain.piece.SlidingPiece;
 
 import java.util.List;
